Dipatcher的handleNewCustomer有问题，是由于capture part route之后指针对应不上导致的，现在改为记录capture part route后insert pos对应的customer id而不是指针，因为customer id是唯一的，该函数还在修改中

+++ src/run/Config.h
@@ -16,7 +16,7 @@ extern int SAMPLE_RATE;
 extern int CUSTOMER_NUM;
 extern int TIME_SLOT_LEN;
 extern int TIME_SLOT_NUM;
-extern int OFF_WORK_TIME; // 骑手收工时间
+extern float OFF_WORK_TIME; // 骑手收工时间
 // R1-R2是store区
 // R3-R4是customer区
 extern float R1;
